+++ Source/UI/Model/LinqToDBModel.cs
@@ -1,0 +1,15 @@
+﻿namespace LinqToDB.LINQPad.UI;
+
+internal sealed class LinqToDBModel : TabModelBase
+{
+	public LinqToDBModel(ConnectionSettings settings)
+		: base(settings)
+	{
+	}
+
+	public bool OptimizeJoins
+	{
+		get => Settings.LinqToDB.OptimizeJoins;
+		set => Settings.LinqToDB.OptimizeJoins = value;
+	}
+}