+++ Source/UI/Settings/SharedConnectionOptions.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace LinqToDB.LINQPad.UI;
+
+internal sealed partial class SharedConnectionOptions
+{
+	public SharedConnectionOptions()
+	{
+		InitializeComponent();
+	}
+}